Refaktoroi: Uudelleenkäytettävät osuudet omaan tiedostoon
