+++ src/main/scala/fi/oph/koski/valpas/rouhinta/ValpasHeturouhintaService.scala
@@ -5,12 +5,9 @@ import fi.oph.koski.db.DatabaseConverters
 import fi.oph.koski.henkilo.{Hetu, OppijaHenkilö}
 import fi.oph.koski.http.HttpStatus
 import fi.oph.koski.log.Logging
-import fi.oph.koski.schema.{Koodistokoodiviite, LocalizedString}
 import fi.oph.koski.util.Timing
-import fi.oph.koski.valpas.opiskeluoikeusrepository.{ValpasHenkilö, ValpasOpiskeluoikeusLaajatTiedot, ValpasOpiskeluoikeusTiedot}
-import fi.oph.koski.valpas.valpasrepository.ValpasOppivelvollisuudenKeskeytys
 import fi.oph.koski.valpas.valpasuser.ValpasSession
-import fi.oph.koski.valpas.{OppijaHakutilanteillaLaajatTiedot, ValpasErrorCategory}
+import fi.oph.koski.valpas.ValpasErrorCategory
 
 import java.time.LocalDate
 
@@ -39,7 +36,7 @@ class ValpasHeturouhintaService(application: KoskiApplication) extends DatabaseC
         .getOppijalista(oppivelvollisetKoskessa.map(_.masterOid))
         .map(oppivelvollisetKoskessa => {
           val (suorittavat, eiSuorittavat) =
-            (oppivelvollisetKoskessa.map(RouhintaOppivelvollinen.apply) ++ oppivelvollisetOnrissa.map(RouhintaOppivelvollinen.apply))
+            (oppivelvollisetKoskessa.map(ValpasRouhintaOppivelvollinen.apply) ++ oppivelvollisetOnrissa.map(ValpasRouhintaOppivelvollinen.apply))
               .partition(_.suorittaaOppivelvollisuutta)
 
           HeturouhinnanTulos(
@@ -86,100 +83,12 @@ class ValpasHeturouhintaService(application: KoskiApplication) extends DatabaseC
 }
 
 case class HeturouhinnanTulos(
-  eiOppivelvollisuuttaSuorittavat: Seq[RouhintaOppivelvollinen],
+  eiOppivelvollisuuttaSuorittavat: Seq[ValpasRouhintaOppivelvollinen],
   oppivelvollisuuttaSuorittavat: Seq[RouhintaPelkkäHetu],
   oppijanumerorekisterinUlkopuoliset: Seq[RouhintaPelkkäHetu],
   oppivelvollisuudenUlkopuoliset: Seq[RouhintaPelkkäHetu],
   virheellisetHetut: Seq[RouhintaPelkkäHetu],
 )
-
-case class RouhintaOppivelvollinen(
-  oppijanumero: ValpasHenkilö.Oid,
-  etunimet: String,
-  sukunimi: String,
-  syntymäaika: Option[LocalDate],
-  hetu: Option[String],
-  viimeisinOppivelvollisuudenSuorittamiseenKelpaavaOpiskeluoikeus: Option[RouhintaOpiskeluoikeus],
-  oppivelvollisuudenKeskeytys: Seq[ValpasOppivelvollisuudenKeskeytys],
-) {
-  def suorittaaOppivelvollisuutta: Boolean =
-    viimeisinOppivelvollisuudenSuorittamiseenKelpaavaOpiskeluoikeus.exists(_.viimeisinTila.koodiarvo == "lasna")
-
-}
-
-object RouhintaOppivelvollinen {
-  def apply(tiedot: OppijaHakutilanteillaLaajatTiedot): RouhintaOppivelvollinen = {
-    val oos = tiedot.oppija.opiskeluoikeudet
-      .filter(_.oppivelvollisuudenSuorittamiseenKelpaava)
-      .flatMap(oo => RouhintaOpiskeluoikeus.apply(oo))
-
-    RouhintaOppivelvollinen(
-      oppijanumero = tiedot.oppija.henkilö.oid,
-      etunimet = tiedot.oppija.henkilö.etunimet,
-      sukunimi = tiedot.oppija.henkilö.sukunimi,
-      syntymäaika = tiedot.oppija.henkilö.syntymäaika,
-      hetu = tiedot.oppija.henkilö.hetu,
-      viimeisinOppivelvollisuudenSuorittamiseenKelpaavaOpiskeluoikeus = oos.sorted.lastOption,
-      oppivelvollisuudenKeskeytys = tiedot.oppivelvollisuudenKeskeytykset.filter(_.voimassa)
-    )
-  }
-
-  def apply(henkilö: OppijaHenkilö): RouhintaOppivelvollinen = RouhintaOppivelvollinen(
-    oppijanumero = henkilö.oid,
-    etunimet = henkilö.etunimet,
-    sukunimi = henkilö.sukunimi,
-    syntymäaika = henkilö.syntymäaika,
-    hetu = henkilö.hetu,
-    viimeisinOppivelvollisuudenSuorittamiseenKelpaavaOpiskeluoikeus = None,
-    oppivelvollisuudenKeskeytys = Nil,
-  )
-}
-
-case class RouhintaOpiskeluoikeus(
-  suorituksenTyyppi: Koodistokoodiviite,
-  päättymispäivä: Option[String],
-  viimeisinTila: Koodistokoodiviite,
-  toimipiste: LocalizedString,
-) extends Ordered[RouhintaOpiskeluoikeus] {
-  override def compare(that: RouhintaOpiskeluoikeus): Int =
-    OrderedOpiskeluoikeusTiedot.compareDates(päättymispäivä, that.päättymispäivä)
-}
-
-object RouhintaOpiskeluoikeus {
-  def apply(oo: ValpasOpiskeluoikeusLaajatTiedot): Option[RouhintaOpiskeluoikeus] = {
-    oo.tarkasteltavaPäätasonSuoritus.flatMap(päätasonSuoritus => {
-      Seq(
-        oo.perusopetusTiedot.map(t => OrderedOpiskeluoikeusTiedot(t)),
-        oo.perusopetuksenJälkeinenTiedot.map(t => OrderedOpiskeluoikeusTiedot(t)),
-      )
-        .flatten
-        .sorted
-        .lastOption
-        .map(viimeisinTila => RouhintaOpiskeluoikeus(
-          suorituksenTyyppi = päätasonSuoritus.suorituksenTyyppi,
-          päättymispäivä = viimeisinTila.tiedot.päättymispäivä,
-          viimeisinTila = viimeisinTila.tiedot.tarkastelupäivänKoskiTila,
-          toimipiste = päätasonSuoritus.toimipiste.nimi,
-        ))
-    })
-  }
-}
-
-case class OrderedOpiskeluoikeusTiedot(
-  tiedot: ValpasOpiskeluoikeusTiedot
-) extends Ordered[OrderedOpiskeluoikeusTiedot] {
-  override def compare(that: OrderedOpiskeluoikeusTiedot): Int =
-    if (tiedot.päättymispäivä != that.tiedot.päättymispäivä) {
-      OrderedOpiskeluoikeusTiedot.compareDates(tiedot.päättymispäivä, that.tiedot.päättymispäivä)
-    } else {
-      OrderedOpiskeluoikeusTiedot.compareDates(tiedot.alkamispäivä, that.tiedot.alkamispäivä)
-    }
-}
-
-object OrderedOpiskeluoikeusTiedot {
-  def compareDates(a: Option[String], b: Option[String]): Int =
-    a.getOrElse("9999-99-99").compare(b.getOrElse("9999-99-99"))
-}
 
 case class RouhintaPelkkäHetu(
   hetu: String,
